+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Dropped support for Python 3.7 and earlier
 - Dropped support for PyQt4 and PySide2
 - Removed unnecessary argument `numPoints` in `QwtSymbol.drawSymbols` and `QwtSymbol.renderSymbols` methods
+- `QwtPlotCanvas`: fixed `BackingStore` feature (`paintAttribute`)
 
 ## Version 0.10.6